backend/requirements.txt
# backend/requirements.txt

fastapi
uvicorn[standard]
python-dotenv
pytesseract
Pillow
google-generativeai
python-multipart